+++ tests/testthat/test-utils.R
@@ -56,4 +56,29 @@ test_that("confl_verb() asks for credentials if it is not set", {
 
   mockery::expect_called(mock_ask2, 1)
   mockery::expect_call(mock_ask2, n = 1, ask_confluence_url())
+
+  # do not cache when options(conflr_dont_cache_envvar) is set FALSE
+  mock_ask3 <- mockery::mock("foo")
+
+  with_mock(
+    "httr::VERB" = mock_success,
+    "conflr::ask_confluence_url" = mock_ask3,
+    "conflr::ask_confluence_username" = mock_ask3,
+    "conflr::ask_confluence_password" = mock_ask3,
+    withr::with_options(
+      list(conflr_dont_cache_envvar = TRUE),
+      withr::with_envvar(
+        list(CONFLUENCE_URL = "", CONFLUENCE_USERNAME = "user", CONFLUENCE_PASSWORD = "pass"), {
+          confl_verb("GET", "/")
+          expect_equal(Sys.getenv("CONFLUENCE_URL"), "")
+          # ensure the option doesn't affect on the envvars that are already set
+          expect_equal(Sys.getenv("CONFLUENCE_USERNAME"), "user")
+          expect_equal(Sys.getenv("CONFLUENCE_PASSWORD"), "pass")
+        }
+      )
+    )
+  )
+
+  mockery::expect_called(mock_ask3, 1)
+  mockery::expect_call(mock_ask3, n = 1, ask_confluence_url())
 })